Fixing weekly totals calculation
Fixing weekly totals calculation which occurs with the -tw flag. Now, t0 is set where new_cases > 0. Weekly totals are now set to calculate values for Sunday prior to t0 with non-overlapping days. If a Sunday prior to t0 is not present in the timeseries dataframe, to allow for weekly sampling to occur starting from a prior Sunday, code handles these cases (NY, for example) by backfilling the frame with zeros to a prior Sunday and using that for the calculation.

Changes to the fixed-t flag were also made so that rather than offsetting stan_data[ts] by days, it offsets by weeks when -tw -ft=1 is used.